refactor(combobox-web): rename combobox to singleselection
